% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{bern_prior_ratio}
\alias{bern_prior_ratio}
\title{Bernoulli prior ratio}
\usage{
bern_prior_ratio(g1, g2, prior_params)
}
\arguments{
\item{g1}{A graph represented as an adjacency matrix.}

\item{g2}{A graph represented as an adjacency matrix.}

\item{prior_params}{A list containing the parameter \code{phi}.}
}
\value{
A numeric value representing the prior ratio.
}
\description{
Computes the prior ratio between two graphs under a Bernoulli prior.
}
\examples{
g1 <- diag(0, 3)
g2 <- matrix(c(0, 1, 0,
               1, 0, 0,
               0, 0, 0), 3, 3)
bern_prior_ratio(g1, g2, list(phi = 0.5))
}
