% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{beta_prior_ratio}
\alias{beta_prior_ratio}
\title{Beta prior ratio}
\usage{
beta_prior_ratio(g1, g2, prior_params)
}
\arguments{
\item{g1}{A graph represented as an adjacency matrix.}

\item{g2}{A graph represented as an adjacency matrix.}

\item{prior_params}{A list with parameters \code{a}, \code{b}, and \code{m}.}
}
\value{
A numeric value representing the prior ratio.
}
\description{
Computes the prior ratio between two graphs under a Beta prior.
}
\examples{
g1 <- diag(0, 3)
g2 <- matrix(c(0, 1, 0,
               1, 0, 0,
               0, 0, 0), 3, 3)
beta_prior_ratio(g1, g2, list(a = 1, b = 1, m = 3))
}
