% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{calculate_shifted_x}
\alias{calculate_shifted_x}
\title{Calculate the Shifted X Matrix}
\usage{
calculate_shifted_x(x, alpha, lambda, f)
}
\arguments{
\item{x}{The data matrix.}

\item{alpha}{The current values of alpha.}

\item{lambda}{The current values of lambda.}

\item{f}{The current factor values.}
}
\value{
The shifted data matrix.
}
\description{
This function computes the shifted X matrix by subtracting alpha and lambda * f from each row of the data.
}
\examples{
set.seed(123)
x <- matrix(rnorm(20), 5, 4)
alpha <- rnorm(4)
lambda <- rnorm(4)
f <- rnorm(5)
calculate_shifted_x(x, alpha, lambda, f)
}
