% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{initialize_chains}
\alias{initialize_chains}
\title{Initialize MCMC chains for lattice model}
\usage{
initialize_chains(
  steps,
  L,
  p,
  n,
  has_start,
  start_lambda,
  start_alpha,
  start_mus,
  start_kw,
  start_f,
  start_delta,
  start_gl,
  start_kl,
  start_nu,
  start_theta,
  start_gamma,
  start_corr_gamma,
  non_continuous,
  z
)
}
\arguments{
\item{steps}{Number of MCMC steps}

\item{L}{Number of groups}

\item{p}{Number of variables}

\item{n}{Vector of sample sizes for each group}

\item{has_start}{Logical indicating if starting values are provided}

\item{start_lambda}{Starting values for lambda parameters}

\item{start_alpha}{Starting values for alpha parameters}

\item{start_mus}{Starting values for mu parameters}

\item{start_kw}{Starting values for Kw matrix}

\item{start_f}{Starting values for factor scores}

\item{start_delta}{Starting values for delta parameters}

\item{start_gl}{Starting values for group graphs}

\item{start_kl}{Starting values for group precision matrices}

\item{start_nu}{Starting values for nu parameters}

\item{start_theta}{Starting values for theta matrix}

\item{start_gamma}{Starting values for gamma matrix}

\item{start_corr_gamma}{Starting values for correlation matrices}

\item{non_continuous}{Logical indicating non-continuous data}

\item{z}{Starting values for latent variables (if non-continuous)}
}
\value{
List containing initialized chains for all parameters
}
\description{
Creates and initializes all the necessary chains for the Gibbs sampler.
}
