% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{initialize_copula_structures}
\alias{initialize_copula_structures}
\title{Initialize copula structures for non-continuous data}
\usage{
initialize_copula_structures(x, L)
}
\arguments{
\item{x}{List of data matrices (one per group)}

\item{L}{Number of groups}
}
\value{
List containing:
\itemize{
\item lessthan - Array indicating values less than breakpoints
\item greaterthan - Array indicating values greater than breakpoints
\item obs_buckets - Matrix indicating bucket membership for each observation
}
}
\description{
Creates data structures needed for copula modeling of non-continuous variables.
}
