% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{initialize_f_temp}
\alias{initialize_f_temp}
\title{Initialize factor scores chain}
\usage{
initialize_f_temp(L, n, start_f, has_start)
}
\arguments{
\item{L}{Number of groups}

\item{n}{Vector of sample sizes}

\item{start_f}{Starting values for factors}

\item{has_start}{Logical indicating if starting values are provided}
}
\value{
Initialized factor scores list
}
\description{
Initializes the chain for factor scores.
}
\keyword{internal}
