% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{initialize_g_chain}
\alias{initialize_g_chain}
\title{Initialize graph structures chain}
\usage{
initialize_g_chain(steps, p, L, start_gl, has_start)
}
\arguments{
\item{steps}{Number of MCMC steps}

\item{p}{Number of variables}

\item{L}{Number of groups}

\item{start_gl}{Starting values for graphs}

\item{has_start}{Logical indicating if starting values are provided}
}
\value{
Initialized graph structures array
}
\description{
Initializes the chain for group graph structures.
}
\keyword{internal}
