% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{initialize_k_chain}
\alias{initialize_k_chain}
\title{Initialize precision matrices chain}
\usage{
initialize_k_chain(steps, p, L, start_kl, has_start)
}
\arguments{
\item{steps}{Number of MCMC steps}

\item{p}{Number of variables}

\item{L}{Number of groups}

\item{start_kl}{Starting values for precision matrices}

\item{has_start}{Logical indicating if starting values are provided}
}
\value{
Initialized precision matrices array
}
\description{
Initializes the chain for group precision matrices.
}
\keyword{internal}
