% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{initialize_kw_chain}
\alias{initialize_kw_chain}
\title{Initialize Kw matrix chain}
\usage{
initialize_kw_chain(steps, L, start_kw, has_start)
}
\arguments{
\item{steps}{Number of MCMC steps}

\item{L}{Number of groups}

\item{start_kw}{Starting values for Kw}

\item{has_start}{Logical indicating if starting values are provided}
}
\value{
Initialized Kw matrix array
}
\description{
Initializes the chain for the spatial precision matrix Kw.
}
\keyword{internal}
