% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{initialize_mu_chain}
\alias{initialize_mu_chain}
\title{Initialize mu parameters chain}
\usage{
initialize_mu_chain(steps, L, start_mus, has_start)
}
\arguments{
\item{steps}{Number of MCMC steps}

\item{L}{Number of groups}

\item{start_mus}{Starting values for mu}

\item{has_start}{Logical indicating if starting values are provided}
}
\value{
Initialized mu parameters matrix
}
\description{
Initializes the chain for mu parameters.
}
\keyword{internal}
