% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{initialize_theta_chain}
\alias{initialize_theta_chain}
\title{Initialize theta matrix chain}
\usage{
initialize_theta_chain(steps, L, start_theta, has_start)
}
\arguments{
\item{steps}{Number of MCMC steps}

\item{L}{Number of groups}

\item{start_theta}{Starting values for theta}

\item{has_start}{Logical indicating if starting values are provided}
}
\value{
Initialized theta matrix array
}
\description{
Initializes the chain for theta matrix.
}
\keyword{internal}
