% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idcbf.R
\name{inv_permute}
\alias{inv_permute}
\title{Inverse Permutation of Permuted Matrix}
\usage{
inv_permute(G, perm)
}
\arguments{
\item{G}{A permuted square matrix.}

\item{perm}{The permutation vector used to permute the matrix.}
}
\value{
The matrix with inverse permutation applied.
}
\description{
Applies the inverse of a previously stored permutation to return matrix to original node order.
}
\keyword{internal}
