% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{is_identifiable}
\alias{is_identifiable}
\title{Identifiability Check for Graphs}
\usage{
is_identifiable(graph, debugFlag = FALSE)
}
\arguments{
\item{graph}{An adjacency matrix representing the graph.}

\item{debugFlag}{A logical flag to plot the original and complement graphs for debugging.}
}
\value{
A logical indicating whether the graph is identifiable (TRUE) or not (FALSE).
}
\description{
This function checks if the graph is identifiable by examining its complement graph.
It looks for bipartite components in the complement graph (which indicates non-identifiability).
}
\examples{
graph <- matrix(c(
  0, 1, 1,
  1, 0, 1,
  1, 1, 0
), 3, 3, byrow = TRUE)
is_identifiable(graph)
}
