% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{mc_hierarchical_prior}
\alias{mc_hierarchical_prior}
\title{MCMC sampler with hierarchical prior}
\usage{
mc_hierarchical_prior(graph, prior_params)
}
\arguments{
\item{graph}{A graph represented as an adjacency matrix.}

\item{prior_params}{A list of prior parameters including \code{p} and \code{iter}.}
}
\value{
A list with the final graph and edge probabilities.
}
\description{
MCMC sampler with hierarchical prior
}
\examples{
g <- diag(0, 3)
mc_hierarchical_prior(g, list(p = 3, iter = 10, m = 3))
}
