% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idcbf.R
\name{N_KU}
\alias{N_KU}
\title{Normalization Constant for G-Wishart Ratio}
\usage{
N_KU(K, D)
}
\arguments{
\item{K}{A precision matrix.}

\item{D}{A symmetric matrix.}
}
\value{
A scalar normalization constant.
}
\description{
Computes a normalization term used in computing acceptance probabilities.
}
\keyword{internal}
