% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{number_to_binary}
\alias{number_to_binary}
\title{Convert Number to Binary Vector}
\usage{
number_to_binary(number, noBits)
}
\arguments{
\item{number}{The number to be converted to binary.}

\item{noBits}{The number of bits to return. If missing, the full binary representation is returned.}
}
\value{
A numeric vector representing the binary representation of the given number.
}
\description{
This function converts a given number into its binary vector representation.
It returns the least significant bits as specified by 'noBits'.
}
\examples{
GSFM:::number_to_binary(5, 3)  # returns c(1, 0, 1)
}
