% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{run_mc_sampler}
\alias{run_mc_sampler}
\title{Generic MCMC sampler}
\usage{
run_mc_sampler(graph, prior_params, prior_ratio_fn)
}
\arguments{
\item{graph}{A graph represented as an adjacency matrix.}

\item{prior_params}{A list of prior parameters including \code{p} and \code{iter}.}

\item{prior_ratio_fn}{A prior ratio function.}
}
\value{
A list with the final graph and edge probabilities.
}
\description{
Runs a Metropolis-Hastings sampler for graphs.
}
\examples{
g <- diag(0, 3)
GSFM:::run_mc_sampler(g, list(p = 3, iter = 10, phi = 0.5), bern_prior_ratio)
}
