% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{sample_bernoulli_prior}
\alias{sample_bernoulli_prior}
\title{Sample graph from Bernoulli prior}
\usage{
sample_bernoulli_prior(graph, prior_params)
}
\arguments{
\item{graph}{Not used.}

\item{prior_params}{A list with \code{p} and \code{phi}.}
}
\value{
A sampled graph as an adjacency matrix.
}
\description{
Samples a random graph under a Bernoulli prior.
}
\examples{
GSFM:::sample_bernoulli_prior(NULL, list(p = 3, phi = 0.5))
}
