% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{sample_unif_prior}
\alias{sample_unif_prior}
\title{Sample graph from uniform prior}
\usage{
sample_unif_prior(graph, prior_params)
}
\arguments{
\item{graph}{Not used.}

\item{prior_params}{A list with parameter \code{p}.}
}
\value{
A sampled graph as an adjacency matrix.
}
\description{
Samples a random graph from the uniform prior.
}
\examples{
GSFM:::sample_unif_prior(NULL, list(p = 3))
}
