% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{save_progress}
\alias{save_progress}
\title{Save progress during MCMC}
\usage{
save_progress(chains, s)
}
\arguments{
\item{chains}{List containing current MCMC chains}

\item{s}{Current MCMC iteration}
}
\description{
Periodically saves the state of the MCMC chains to disk.
}
