% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{set_hyperparameters}
\alias{set_hyperparameters}
\title{Set hyperparameters for lattice model}
\usage{
set_hyperparameters(p, L, W)
}
\arguments{
\item{p}{Number of variables}

\item{L}{Number of groups}

\item{W}{Spatial adjacency matrix}
}
\value{
List containing all hyperparameters
}
\description{
Defines the default hyperparameters used in the Gibbs sampler.
}
