% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{size}
\alias{size}
\title{Compute the number of edges in an undirected graph}
\usage{
size(g)
}
\arguments{
\item{g}{An undirected graph represented as a square adjacency matrix.}
}
\value{
The number of edges in the graph.
}
\description{
This function computes the total number of edges in an undirected graph represented as an adjacency matrix.
It counts the number of 1s in the upper triangle of the matrix (excluding the diagonal), which represents the edges.
}
\examples{
g <- matrix(c(0, 1, 0,
              1, 0, 1,
              0, 1, 0), nrow = 3, byrow = TRUE)
GSFM:::size(g)
}
