% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{unif_prior_ratio}
\alias{unif_prior_ratio}
\title{Uniform prior ratio}
\usage{
unif_prior_ratio(g1, g2, prior_params)
}
\arguments{
\item{g1}{A graph represented as an adjacency matrix.}

\item{g2}{A graph represented as an adjacency matrix.}

\item{prior_params}{A list of prior parameters (not used in this case).}
}
\value{
A constant value of 1.
}
\description{
Computes the prior ratio between two graphs under a uniform prior. Always returns 1.
}
\examples{
g1 <- diag(0, 3)
g2 <- matrix(c(0, 1, 0,
               1, 0, 0,
               0, 0, 0), 3, 3)
unif_prior_ratio(g1, g2, list())
}
