% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_alphas}
\alias{update_alphas}
\title{Update the Alphas}
\usage{
update_alphas(K, x, n, n0, f, lambda)
}
\arguments{
\item{K}{The precision matrix.}

\item{x}{The data matrix.}

\item{n}{The number of observations.}

\item{n0}{The prior parameter for alpha.}

\item{f}{The current factor values.}

\item{lambda}{The current values of lambda.}
}
\value{
A numeric vector of updated alphas.
}
\description{
This function updates the alphas (alpha) based on the current values of lambda and f.
}
\examples{
set.seed(123)
x <- matrix(rnorm(20), 5, 4)
lambda <- rnorm(4)
f <- rnorm(5)
K <- diag(4)
update_alphas(K, x, 5, 1, f, lambda)
}
