% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{update_chains}
\alias{update_chains}
\title{Update MCMC chains with new results}
\usage{
update_chains(chains, results, L, s)
}
\arguments{
\item{chains}{List containing current MCMC chains}

\item{results}{List of results from parallel updates}

\item{L}{Number of groups}

\item{s}{Current MCMC iteration}
}
\value{
Updated chains list
}
\description{
Incorporates the results from parallel group updates into the main chains.
}
