% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_delta}
\alias{update_delta}
\title{Update Delta}
\usage{
update_delta(c, d, lambda, K)
}
\arguments{
\item{c}{The shape parameter for the inverse gamma prior.}

\item{d}{The scale parameter for the inverse gamma prior.}

\item{lambda}{The current values of lambda.}

\item{K}{The precision matrix.}
}
\value{
A scalar representing the updated delta value.
}
\description{
This function updates the delta parameter based on the current values of lambda and the precision matrix K.
}
\examples{
set.seed(123)
lambda <- rnorm(4)
K <- diag(4)
update_delta(2, 1, lambda, K)
}
