% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_factors}
\alias{update_factors}
\title{Update the Factors}
\usage{
update_factors(K, x, lambda, alpha)
}
\arguments{
\item{K}{The precision matrix.}

\item{x}{The data matrix with rows representing observations.}

\item{lambda}{The current values of lambda.}

\item{alpha}{The current values of alpha.}
}
\value{
A numeric vector of updated factors (f).
}
\description{
This function updates the factors (f) based on the current values of lambda and alpha.
}
\examples{
set.seed(123)
x <- matrix(rnorm(20), 5, 4)
lambda <- rnorm(4)
alpha <- rnorm(4)
K <- diag(4)
update_factors(K, x, lambda, alpha)
}
