% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{update_groups_parallel}
\alias{update_groups_parallel}
\title{Update group parameters in parallel}
\usage{
update_groups_parallel(
  cl,
  L,
  chains,
  x,
  n,
  p,
  params,
  non_continuous,
  copula,
  copula_structs,
  s
)
}
\arguments{
\item{cl}{Parallel cluster object}

\item{L}{Number of groups}

\item{chains}{List containing current MCMC chains}

\item{x}{List of data matrices}

\item{n}{Vector of sample sizes}

\item{p}{Number of variables}

\item{params}{List of hyperparameters}

\item{non_continuous}{Logical indicating non-continuous data}

\item{copula}{Logical indicating copula model}

\item{copula_structs}{List of copula structures}

\item{s}{Current MCMC iteration}
}
\value{
List of updated parameters for each group
}
\description{
Distributes the group parameter updates across multiple cores.
}
