% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_kw}
\alias{update_kw}
\title{Update Spatial Precision Matrix}
\usage{
update_kw(deltaw, rho, W, mus)
}
\arguments{
\item{deltaw}{Degrees of freedom for G-Wishart prior}

\item{rho}{Spatial autocorrelation parameter}

\item{W}{Adjacency matrix for spatial structure (L x L)}

\item{mus}{Current mean parameters (vector of length L)}
}
\value{
Updated spatial precision matrix (L x L)
}
\description{
Updates the spatial precision matrix Kw using a G-Wishart distribution.
Ensures numerical symmetry and handles single-group case.
}
