% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_lambdas_spikeslab}
\alias{update_lambdas_spikeslab}
\title{Update Lambda Values with Spike-and-Slab Prior}
\usage{
update_lambdas_spikeslab(K, x, f, delta, alpha, lambda, pi_lambda)
}
\arguments{
\item{K}{The precision matrix.}

\item{x}{The data matrix.}

\item{f}{The current factor values.}

\item{delta}{The regularization parameter.}

\item{alpha}{The current values of alpha.}

\item{lambda}{The current values of lambda.}

\item{pi_lambda}{The probability of lambda being non-zero.}
}
\value{
A matrix of updated lambda values.
}
\description{
Update Lambda Values with Spike-and-Slab Prior
}
\examples{
set.seed(123)
x <- matrix(rnorm(20), 5, 4)
f <- rnorm(5)
alpha <- rnorm(4)
lambda <- rnorm(4)
K <- diag(4)
update_lambdas_spikeslab(K, x, f, 1, alpha, lambda, 0.8)
}
