% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_pi_prob}
\alias{update_pi_prob}
\title{Update Probability of Lambda Being Non-Zero}
\usage{
update_pi_prob(lambda)
}
\arguments{
\item{lambda}{The current values of lambda.}
}
\value{
A scalar representing the updated probability of lambda being non-zero.
}
\description{
Update Probability of Lambda Being Non-Zero
}
\examples{
set.seed(123)
lambda <- c(0, 1.2, 0, 0.8)
update_pi_prob(lambda)
}
