% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler_helpers.R
\name{update_shared_parameters}
\alias{update_shared_parameters}
\title{Update shared parameters}
\usage{
update_shared_parameters(chains, params, s)
}
\arguments{
\item{chains}{List containing current MCMC chains}

\item{params}{List of hyperparameters}

\item{s}{Current MCMC iteration}
}
\value{
Updated chains list
}
\description{
Updates parameters that are shared across groups (theta, gamma, nu).
}
