% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_zs}
\alias{update_zs}
\title{Update the Z Values}
\usage{
update_zs(
  K,
  G,
  x,
  lambda,
  alpha,
  f,
  z,
  copula = FALSE,
  lessthan_buckets,
  greaterthan_buckets,
  buckets
)
}
\arguments{
\item{K}{The precision matrix.}

\item{G}{The group indicator matrix.}

\item{x}{The data matrix.}

\item{lambda}{The current values of lambda.}

\item{alpha}{The current values of alpha.}

\item{f}{The current factor values.}

\item{z}{The current values of z.}

\item{copula}{A logical value indicating if copula sampling should be used.}

\item{lessthan_buckets}{A matrix of indices for values less than the bucket.}

\item{greaterthan_buckets}{A matrix of indices for values greater than the bucket.}

\item{buckets}{A matrix of bucket indices for each observation.}
}
\value{
A matrix of updated z values.
}
\description{
This function updates the latent variable z using truncated normal or copula-based sampling.
}
\examples{
set.seed(123)
n <- 5; p <- 4
K <- diag(p)
G <- matrix(1, p, p)
x <- matrix(rbinom(n * p, 1, 0.5), n, p)
lambda <- rnorm(p)
alpha <- rnorm(p)
f <- rnorm(n)
z <- matrix(rnorm(n * p), n, p)
buckets <- matrix(sample(1:3, n * p, TRUE), n, p)
lessthan_buckets <- array(TRUE, dim = c(n, p, 3))
greaterthan_buckets <- array(TRUE, dim = c(n, p, 3))
update_zs(K, G, x, lambda, alpha, f, z, FALSE, lessthan_buckets, greaterthan_buckets, buckets)
}
